Add files via upload
The Database_MAF_Script.sql when executed will create the Dabase MAF on the instance. 
While the Data_Insertion_script.sql is the generated script used to populate the database. 
Tables_Creation_Script is used to create all schemas in the database. 
Testing_Scripts.sql are written script to test the relationship between each schema.